get rid of a few warnings

+++ SourceX/DiabloUI/scrollbar.h
@@ -8,7 +8,7 @@ namespace dvl {
 extern Art ArtScrollBarBackground;
 extern Art ArtScrollBarThumb;
 extern Art ArtScrollBarArrow;
-const Uint16 SCROLLBAR_BG_WIDTH = 25;
+const int SCROLLBAR_BG_WIDTH = 25;
 
 extern Art ArtScrollBarArrow;
 enum ScrollBarArrowFrame {
@@ -19,7 +19,7 @@ enum ScrollBarArrowFrame {
 };
 
 extern Art ArtScrollBarThumb;
-const Uint16 SCROLLBAR_ARROW_WIDTH = 25;
+const int SCROLLBAR_ARROW_WIDTH = 25;
 
 inline SDL_Rect UpArrowRect(const UiScrollBar *sb)
 {
